+++ src/main/java/org/opensearch/security/privileges/PrivilegesInterceptor.java
@@ -32,7 +32,6 @@ import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.security.resolver.IndexResolverReplacer.Resolved;
-import org.opensearch.security.securityconf.DynamicConfigModel;
 import org.opensearch.security.user.User;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.client.Client;
@@ -80,10 +79,8 @@ public class PrivilegesInterceptor {
         final ActionRequest request,
         final String action,
         final User user,
-        final DynamicConfigModel config,
         final Resolved requestedResolved,
-        final PrivilegesEvaluationContext context,
-        final TenantPrivileges tenantPrivileges
+        final PrivilegesEvaluationContext context
     ) {
         throw new RuntimeException("not implemented");
     }